+++ src/api/Factory.hpp
@@ -39,12 +39,13 @@ class Factory
     /** \brief Creates an API instance based on the given API type.
      *
      * \param api     type of the API instance to create
+     * \param planWb  current pricing plan for Weatherbit
      * \param planWs  current pricing plan for Weatherstack
      * \param key     the API key for that instance (optional)
      * \return Returns a unique_ptr to the created instance.
      *         Returns nullptr, if api is ApiType::none or not supported.
      */
-    static std::unique_ptr<API> create(const ApiType api, const PlanWeatherstack planWs, const std::string& key = std::string());
+    static std::unique_ptr<API> create(const ApiType api, const PlanWeatherbit planWb, const PlanWeatherstack planWs, const std::string& key = std::string());
 }; // class
 
 } // namespace